Feat: Extract the save buttons for dataset and chat configurations to separate files to increase permission control #3221
